refactor(run): extract forge_run submodule (Wave C5)
Sixth and largest cut of the run.rs split (2,388 -> ~630). Moves the entire
forge mission pipeline into run/forge_run.rs (~1,800 lines): config knobs,
run_forge_mission (the Coder->Verifier fix loop), Planner/Verifier role turns,
parse_verifier_response, round-history scoring (best_round / score_declining),
EphemeralMissionGuard, capture_git_diff/base_sha, short_sha, git_reset_hard,
forge_phase0_preflight, truncate_diff_for_review, forge_confirm_submit, plus all
of their unit tests (verifier-parse, best-round, score-declining, short_sha,
human-review, build/test-gate) and the git/round/restore_env test helpers.

A 3-region move around three shared stays kept in run.rs: env_flag_enabled,
agent_auto_approve_enabled, extract_assistant_text (forge_run reaches them via
use super::*). Re-exports: pub run_forge_mission (lib.rs/main.rs) + pub(crate)
forge_auto_approve_enabled (the C4 runtime builders consult it). The
forge_model_map test stays in run.rs (tests forge::types, not a moved fn).

Behavior-identical pure code movement (no A/B). whole-crate build green; fmt +
clippy (lean + --all-features) -D warnings clean; cargo test --lib 1035 passed
(unchanged — tests relocated, not dropped). Stacked on #149 (C6). Wave C is now
6/6 (C1-C6); only C-final (drop the crate-wide too_many_lines/too_many_arguments
allows) remains.
